Add console app that creates many file watchers
HttpClient is used to check networking is working after each file
watcher is created. Eventually the http request will fail with an
error about the socket not being connected.

HttpClient request failed. An error occurred while sending the request
Error: SecureChannelFailure (Unable to write data to the transport connection: The socket is not connected.)
Watchers created: 254